Added utility procedures and audio_global from previous godot project
We're going to add one to two more features and some API usability improvements and this should be a sufficient enough skeleton to survive 7DRL. I may or may not add projectiles here, but that might be too much to add.

There is no inventory system since that's kind of game specific and the rest of the stuff should be coverable by Godot's built in nodes (mostly UI input stuff).