setting up horde mode room
we're going to do dungeon generation with no zombies today. I'm probably going to write some utility functions for the Director's AI (such as knowing if we can see another room or not...).

Also note to self: The Safe Room is always going to be the last room using the decided dungeon generation scheme.