director is working as intended
at least for hordes, need to introduce spawn cap.

Once that's done I can move on to the rest of the things. As while the director isn't done, he should be in good enough shape that I can mold the other things I need, and then come back. It helps that the director isn't a trainwreck.